refactor: move SpectacularAppComponent and related Angular modules to application-testing
