+++ packages/spectacular/src/index.ts
@@ -1,6 +1,7 @@
 // Application testing
-export * from './lib/application-testing/bootstrap-spectacular-application';
-export * from './lib/application-testing/bootstrap-spectacular-application-options';
+export * from './lib/application-testing/bootstrapping/bootstrap-spectacular-application';
+export * from './lib/application-testing/bootstrapping/bootstrap-spectacular-application-options';
+export * from './lib/application-testing/app-component/spectacular-app.component';
 
 // Feature testing
 //  Feature testing module
@@ -18,6 +19,3 @@ export * from './lib/feature-testing/testbed/spectacular-feature-testbed';
 export * from './lib/pipe-test-harness/create-pipe-test-harness';
 export * from './lib/pipe-test-harness/create-pipe-test-harness-options';
 export * from './lib/pipe-test-harness/pipe-test-harness';
-
-// Root level
-export * from './lib/root-level/spectacular-app.component';
