+++ packages/spectacular/src/lib/application-testing/app-component/spectacular-app.component.ts
