+++ packages/spectacular/src/lib/application-testing/bootstrapping/bootstrap-spectacular-application.ts
